+++ bindings/python/src/encoding.rs
@@ -115,7 +115,7 @@ impl PyEncoding {
     /// Returns:
     ///     :obj:`List[int]`: The list of IDs
     #[getter]
-    fn get_ids(&self) -> Vec<u32> {
+    fn get_ids(&self) -> Vec<u64> {
         self.encoding.get_ids().to_vec()
     }
 
@@ -147,7 +147,7 @@ impl PyEncoding {
     /// Returns:
     ///     A :obj:`List` of :obj:`Optional[int]`: A list of optional word index.
     #[getter]
-    fn get_words(&self, py: Python<'_>) -> PyResult<Vec<Option<u32>>> {
+    fn get_words(&self, py: Python<'_>) -> PyResult<Vec<Option<u64>>> {
         deprecation_warning(
             py,
             "0.9.4",
@@ -169,7 +169,7 @@ impl PyEncoding {
     /// Returns:
     ///     A :obj:`List` of :obj:`Optional[int]`: A list of optional word index.
     #[getter]
-    fn get_word_ids(&self) -> Vec<Option<u32>> {
+    fn get_word_ids(&self) -> Vec<Option<u64>> {
         self.encoding.get_word_ids().to_vec()
     }
 
@@ -194,7 +194,7 @@ impl PyEncoding {
     /// Returns:
     ///     :obj:`List[int]`: The list of type ids
     #[getter]
-    fn get_type_ids(&self) -> Vec<u32> {
+    fn get_type_ids(&self) -> Vec<u64> {
         self.encoding.get_type_ids().to_vec()
     }
 
@@ -217,7 +217,7 @@ impl PyEncoding {
     /// Returns:
     ///     :obj:`List[int]`: The special tokens mask
     #[getter]
-    fn get_special_tokens_mask(&self) -> Vec<u32> {
+    fn get_special_tokens_mask(&self) -> Vec<u64> {
         self.encoding.get_special_tokens_mask().to_vec()
     }
 
@@ -230,7 +230,7 @@ impl PyEncoding {
     /// Returns:
     ///    :obj:`List[int]`: The attention mask
     #[getter]
-    fn get_attention_mask(&self) -> Vec<u32> {
+    fn get_attention_mask(&self) -> Vec<u64> {
         self.encoding.get_attention_mask().to_vec()
     }
 
@@ -266,7 +266,7 @@ impl PyEncoding {
     ///     :obj:`Tuple[int, int]`: The range of tokens: :obj:`(first, last + 1)`
     #[pyo3(signature = (word_index, sequence_index = 0))]
     #[pyo3(text_signature = "(self, word_index, sequence_index=0)")]
-    fn word_to_tokens(&self, word_index: u32, sequence_index: usize) -> Option<(usize, usize)> {
+    fn word_to_tokens(&self, word_index: u64, sequence_index: usize) -> Option<(usize, usize)> {
         self.encoding.word_to_tokens(word_index, sequence_index)
     }
 
@@ -282,7 +282,7 @@ impl PyEncoding {
     ///     :obj:`Tuple[int, int]`: The range of characters (span) :obj:`(first, last + 1)`
     #[pyo3(signature = (word_index, sequence_index = 0))]
     #[pyo3(text_signature = "(self, word_index, sequence_index=0)")]
-    fn word_to_chars(&self, word_index: u32, sequence_index: usize) -> Option<Offsets> {
+    fn word_to_chars(&self, word_index: u64, sequence_index: usize) -> Option<Offsets> {
         self.encoding.word_to_chars(word_index, sequence_index)
     }
 
@@ -333,7 +333,7 @@ impl PyEncoding {
     /// Returns:
     ///     :obj:`int`: The index of the word in the relevant input sequence.
     #[pyo3(text_signature = "(self, token_index)")]
-    fn token_to_word(&self, token_index: usize) -> Option<u32> {
+    fn token_to_word(&self, token_index: usize) -> Option<u64> {
         let (_, word_idx) = self.encoding.token_to_word(token_index)?;
         Some(word_idx)
     }
@@ -366,7 +366,7 @@ impl PyEncoding {
     ///     :obj:`int`: The index of the word that contains this char in the input sequence
     #[pyo3(signature = (char_pos, sequence_index = 0))]
     #[pyo3(text_signature = "(self, char_pos, sequence_index=0)")]
-    fn char_to_word(&self, char_pos: usize, sequence_index: usize) -> Option<u32> {
+    fn char_to_word(&self, char_pos: usize, sequence_index: usize) -> Option<u64> {
         self.encoding.char_to_word(char_pos, sequence_index)
     }
